fix: page not loading when navbar links are clicked
