+++ src/components/ui/Navbar.tsx
@@ -2,18 +2,7 @@
 
 import Image from "next/image";
 import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { Fragment, Suspense, useCallback, useEffect, useState } from "react";
-
-export default function Navbar() {
-  // Adding suspense here
-  // Ref: https://nextjs.org/docs/messages/missing-suspense-with-csr-bailout
-  return (
-    <Suspense>
-      <NavbarContent />
-    </Suspense>
-  );
-}
+import { Fragment, useCallback, useEffect, useState } from "react";
 
 const NAV_LINKS = [
   { href: "/#projects", label: "Projects" },
@@ -23,23 +12,7 @@ const NAV_LINKS = [
   { href: "/contact", label: "Contact" }
 ];
 
-export function NavbarContent() {
-  const pathname = usePathname();
-
-  useEffect(() => {
-    // window.location.hash has the id to scroll to
-    const hash = window.location.hash?.substring(1); // remove the '#'
-
-    if (hash) {
-      const element = document.getElementById(hash);
-      if (element) {
-        setTimeout(() => {
-          element.scrollIntoView({ behavior: "smooth" });
-        }, 1000);
-      }
-    }
-  }, [pathname]);
-
+export default function Navbar() {
   return (
     <nav className="fixed top-0 left-0 z-1 w-full border-b-1 border-neutral-600 bg-neutral-900/30 backdrop-blur-2xl">
       <div className="container">
